regime-trader/requirements.txt
# ─── Pinned versions for cross-platform reproducibility ─────────────────────
# These exact pins make Linux backtest results deterministic.
# Windows native Python (MSVC/ILP64 OpenBLAS) will still differ from Linux
# (GCC/LP64 OpenBLAS) due to different floating-point accumulation in HMM EM.
# Canonical platform: Linux (native or WSL2 Ubuntu-24.04).
# Regenerate: pip freeze > requirements.txt  (inside activated venv)

# ── HMM & ML ─────────────────────────────────────────────────────────────────
hmmlearn==0.3.3
scikit-learn==1.8.0
scipy==1.17.1
numpy==2.4.4
joblib==1.5.3
threadpoolctl==3.6.0

# ── Broker / market data ─────────────────────────────────────────────────────
alpaca-py==0.43.2
alpaca-trade-api==3.2.0
aiohttp==3.13.5
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
frozenlist==1.8.0
multidict==6.7.1
propcache==0.4.1
yarl==1.23.0
websockets==10.4
websocket-client==1.9.0
sseclient-py==1.9.0
requests==2.33.1
urllib3==1.26.20

# ── Data processing ──────────────────────────────────────────────────────────
pandas==3.0.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0

# ── Technical analysis ───────────────────────────────────────────────────────
ta==0.11.0

# ── Config & environment ─────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.2.2

# ── Pydantic (required by alpaca-py) ─────────────────────────────────────────
pydantic==2.13.3
pydantic_core==2.46.3
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ── Scheduling ───────────────────────────────────────────────────────────────
schedule==1.2.2

# ── Terminal dashboard ────────────────────────────────────────────────────────
rich==15.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.20.0

# ── Network / HTTP utils ─────────────────────────────────────────────────────
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
attrs==26.1.0
msgpack==1.0.3
deprecation==2.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==9.0.3
pluggy==1.6.0
iniconfig==2.3.0
packaging==26.1
